+++ components-starters/components-starter-object-storage/src/main/java/com/skrstop/framework/components/starter/objectStorage/service/impl/CosObjectStorageServiceImpl.java
@@ -1,6 +1,8 @@
 package com.skrstop.framework.components.starter.objectStorage.service.impl;
 
 import cn.hutool.core.map.MapUtil;
+import cn.hutool.core.util.URLUtil;
+import com.alibaba.fastjson2.JSON;
 import com.qcloud.cos.COSClient;
 import com.qcloud.cos.ClientConfig;
 import com.qcloud.cos.auth.BasicCOSCredentials;
@@ -13,6 +15,7 @@ import com.qcloud.cos.region.Region;
 import com.qcloud.cos.transfer.Download;
 import com.qcloud.cos.transfer.TransferManager;
 import com.qcloud.cos.transfer.Upload;
+import com.qcloud.cos.utils.UrlEncoderUtils;
 import com.skrstop.framework.components.starter.objectStorage.configuration.CosProperties;
 import com.skrstop.framework.components.starter.objectStorage.entity.CosStorageTemplateSign;
 import com.skrstop.framework.components.starter.objectStorage.entity.StorageTemplateSign;
@@ -61,6 +64,12 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
 
     public CosObjectStorageServiceImpl(CosProperties cosProperties) {
         this.cosProperties = cosProperties;
+        if (StrUtil.isNotBlank(cosProperties.getBasePath())) {
+            basePath = cosProperties.getBasePath();
+        }
+        if (StrUtil.isBlank(basePath)) {
+            basePath = "/";
+        }
         try {
             // 构建client
             COSCredentials cred = new BasicCOSCredentials(cosProperties.getSecretId(), cosProperties.getSecretKey());
@@ -84,12 +93,6 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
             this.transferManager = new TransferManager(this.cosClient, executor);
         } catch (Exception e) {
             throw new IllegalArgumentException("cos客户端创建失败");
-        }
-        if (StrUtil.isNotBlank(cosProperties.getBasePath())) {
-            basePath = cosProperties.getBasePath();
-        }
-        if (StrUtil.isBlank(basePath)) {
-            basePath = "/";
         }
     }
 
@@ -252,18 +255,45 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
     }
 
     @Override
+    public String getPublicAccessUrl(String bucketName, String targetPath) {
+        bucketName = this.getOrDefaultBucketName(bucketName);
+        if (StrUtil.isNotBlank(cosProperties.getAccessUrlHost())) {
+            return this.cosProperties.getAccessUrlProtocol() + "://" + this.cosProperties.getAccessUrlHost() + targetPath;
+        } else {
+            StringBuilder strBuilder = new StringBuilder();
+            strBuilder.append(this.cosClient.getClientConfig().getHttpProtocol().toString()).append("://");
+            strBuilder.append(this.cosClient.getClientConfig().getEndpointBuilder()
+                    .buildGeneralApiEndpoint(bucketName));
+            strBuilder.append(UrlEncoderUtils.encodeUrlPath(targetPath));
+            return strBuilder.toString();
+        }
+    }
+
+    @Override
+    public Map<String, String> getPublicAccessUrl(String bucketName, List<String> targetPath) {
+        bucketName = this.getOrDefaultBucketName(bucketName);
+        Map<String, String> result = new HashMap<>(targetPath.size(), 1);
+        for (String path : targetPath) {
+            result.put(path, this.getPublicAccessUrl(bucketName, path));
+        }
+        return result;
+    }
+
+    @Override
     public String getTemporaryAccessUrl(String bucketName, String targetPath, long expireTime) {
         bucketName = this.getOrDefaultBucketName(bucketName);
         LocalDateTime endTime = LocalDateTime.now().plusSeconds(expireTime);
-        URL url = this.cosClient.generatePresignedUrl(bucketName, targetPath, DateUtil.toDate(endTime), HttpMethodName.GET);
+        URL url = this.cosClient.generatePresignedUrl(bucketName, targetPath
+                , DateUtil.toDate(endTime), HttpMethodName.GET
+                , new HashMap<String, String>(), new HashMap<String, String>(), false, false);
         try {
             URI newUrl = new URI(StrUtil.blankToDefault(cosProperties.getAccessUrlProtocol(), url.getProtocol())
                     , url.getUserInfo()
                     , StrUtil.blankToDefault(cosProperties.getAccessUrlHost(), url.getHost())
                     , url.getPort()
                     , url.getPath()
-                    , url.getQuery()
-                    , ""
+                    , URLUtil.decode(url.getQuery())
+                    , null
             );
             String result = newUrl.toString();
             result = result.replace(":80", "");
@@ -285,7 +315,7 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
     }
 
     @Override
-    public <T extends StorageTemplateSign> T getTemporaryAccessSign(String bucketName, String targetPath, long expireSecondTime, Long minSize, Long maxSize, List<ContentTypeEnum> contentType) {
+    public <T extends StorageTemplateSign> T getTemporaryUploadSign(String bucketName, String targetPath, long expireSecondTime, Long minSize, Long maxSize, List<ContentTypeEnum> contentType) {
         bucketName = this.getOrDefaultBucketName(bucketName);
         CosStorageTemplateSign sign = new CosStorageTemplateSign();
         try {
@@ -293,15 +323,19 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
             config.put("secretId", cosProperties.getSecretId());
             config.put("secretKey", cosProperties.getSecretKey());
             // 临时密钥有效时长，单位是秒
-            config.put("durationSeconds", expireSecondTime);
+            config.put("durationSeconds", Math.toIntExact(expireSecondTime));
             // 换成你的 bucket
             config.put("bucket", bucketName);
             // 换成 bucket 所在地区
             config.put("region", cosProperties.getRegion());
-            // 可以通过 allowPrefixes 指定前缀数组, 例子： a.jpg 或者 a/* 或者 * (使用通配符*存在重大安全风险, 请谨慎评估使用)
-            config.put("allowPrefixes", CollectionUtil.newArrayList(targetPath));
+            // policy
+            Map<String, Object> policy = new HashMap<>();
+            policy.put("version", "2.0");
+            Map<String, Object> statement = new HashMap<>();
+            policy.put("statement", CollectionUtil.newArrayList(statement));
+            statement.put("effect", "allow");
             // 密钥的权限列表。简单上传和分片需要以下的权限，其他权限列表请看 https://cloud.tencent.com/document/product/436/31923
-            config.put("allowActions", CollectionUtil.newArrayList(
+            String[] actions = {
                     // 简单上传
                     "name/cos:PutObject",
                     "name/cos:PostObject",
@@ -310,33 +344,54 @@ public class CosObjectStorageServiceImpl implements ObjectStorageService {
                     "name/cos:ListMultipartUploads",
                     "name/cos:ListParts",
                     "name/cos:UploadPart",
-                    "name/cos:CompleteMultipartUpload"
+                    "name/cos:CompleteMultipartUpload",
+                    "name/cos:AbortMultipartUpload"
+            };
+            statement.put("action", actions);
+            // 可以通过 allowPrefixes 指定前缀数组, 例子： a.jpg 或者 a/* 或者 * (使用通配符*存在重大安全风险, 请谨慎评估使用)
+            if (!targetPath.startsWith("/")) {
+                targetPath = "/" + targetPath;
+            }
+            if (StrUtil.isNotBlank(this.basePath)
+                    && !"/".equals(this.basePath)) {
+                targetPath = this.basePath + targetPath;
+            }
+            statement.put("resource", CollectionUtil.newArrayList(
+                    String.format("qcs::cos:%s:uid/%s:%s%s",
+                            cosProperties.getRegion()
+                            , bucketName.substring(bucketName.lastIndexOf("-") + 1)
+                            , bucketName
+                            , targetPath)
             ));
-            Map<String, Object> policy = new HashMap<>();
-            Map<String, Object> policyCondition = new HashMap<>();
-            policy.put("condition", policyCondition);
+            Map<String, Object> conditions = new HashMap<>();
             if (ObjectUtil.isNotNull(minSize)) {
                 // 限制大小
-                policyCondition.put("numeric_greater_than_equal", MapUtil.builder("cos:content-length", minSize * 1024));
+                Map<String, ArrayList<Long>> val = MapUtil.builder("cos:content-length", CollectionUtil.newArrayList(minSize)).map();
+                conditions.put("numeric_greater_than_equal", val);
             }
             if (ObjectUtil.isNotNull(maxSize)) {
                 // 限制大小
-                policyCondition.put("numeric_less_than_equal", MapUtil.builder("cos:content-length", maxSize * 1024));
+                Map<String, ArrayList<Long>> val = MapUtil.builder("cos:content-length", CollectionUtil.newArrayList(maxSize)).map();
+                conditions.put("numeric_less_than_equal", val);
             }
             if (CollectionUtil.isNotEmpty(contentType)) {
                 // 限制类型
-                List<String> contentTypeStrList = contentType.stream().map(ContentTypeEnum::getContentType).collect(Collectors.toList());
-                policyCondition.put("string_like", MapUtil.builder("cos:content-type", contentTypeStrList));
+                Set<String> list = contentType.stream().map(ContentTypeEnum::getContentType).collect(Collectors.toSet());
+                Map<String, Set<String>> val = MapUtil.builder("cos:content-type", list).map();
+                conditions.put("string_like", val);
             }
-            if (MapUtil.isNotEmpty(policyCondition)) {
-                config.put("policy", policy);
-            }
+            statement.put("condition", conditions);
+            config.put("policy", JSON.toJSONString(policy));
             Response response = CosStsClient.getCredential(config);
             sign.setTmpSecretId(response.credentials.tmpSecretId);
             sign.setTmpSecretKey(response.credentials.tmpSecretKey);
             sign.setSessionToken(response.credentials.sessionToken);
+            sign.setToken(response.credentials.token);
             sign.setExpireSecondTime(expireSecondTime);
             sign.setExpireDateTime(LocalDateTime.now().plusSeconds(expireSecondTime));
+            sign.setTargetPath(targetPath);
+            sign.setBucketName(bucketName);
+            sign.setRegion(this.cosProperties.getRegion());
             return (T) sign;
         } catch (Exception e) {
             log.error(e.getMessage(), e);