+++ components-starters/components-starter-object-storage/src/main/java/com/skrstop/framework/components/starter/objectStorage/service/impl/DynamicObjectStorageServiceImpl.java
@@ -101,13 +101,23 @@ public class DynamicObjectStorageServiceImpl implements DynamicObjectStorageServ
     }
 
     @Override
+    public String getPublicAccessUrl(String bucketName, String targetPath) {
+        return this.objectStorageService.getPublicAccessUrl(bucketName, targetPath);
+    }
+
+    @Override
+    public Map<String, String> getPublicAccessUrl(String bucketName, List<String> targetPath) {
+        return this.objectStorageService.getPublicAccessUrl(bucketName, targetPath);
+    }
+
+    @Override
     public Map<String, String> getTemporaryAccessUrl(String dsKey, String bucketName, List<String> targetPath, long expireTime) {
         return this.objectStorageService.getTemporaryAccessUrl(bucketName, targetPath, expireTime);
     }
 
     @Override
     public <T extends StorageTemplateSign> T getTemporaryAccessSign(String dsKey, String bucketName, String targetPath, long expireSecondTime, Long minSize, Long maxSize, List<ContentTypeEnum> contentType) {
-        return this.objectStorageService.getTemporaryAccessSign(bucketName, targetPath, expireSecondTime, minSize, maxSize, contentType);
+        return this.objectStorageService.getTemporaryUploadSign(bucketName, targetPath, expireSecondTime, minSize, maxSize, contentType);
     }
 
     @Override